перенос

+++ ewm-main-service/mainServer/src/main/java/ru/practicum/dto/EventFullDto.java
@@ -13,7 +13,7 @@ public class EventFullDto {
     private Long id;
     private String annotation; //описание
     private CategoryDto category; //Категория
-    private Integer confirmedRequests;//Количество одобренных заявок на участие в данном событии
+    private Long confirmedRequests;//Количество одобренных заявок на участие в данном событии
     private String createdOn;//Дата и время создания события (в формате "yyyy-MM-dd HH:mm:ss")
     private String description;//Полное описание события
     private String eventDate;//Дата и время на которые намечено событие (в формате "yyyy-MM-dd HH:mm:ss")
@@ -27,5 +27,5 @@ public class EventFullDto {
     private Boolean requestModeration;//Нужна ли пре-модерация заявок на участие
     private String state; //Список состояний жизненного цикла события
     private String title;//Заголовок
-    private Integer views;//Количество просмотрев события
+    private Long views;//Количество просмотрев события
 }
